requirements.txt
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cycler==0.10.0
easydict==1.9
kiwisolver==1.2.0
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.1
olefile==0.46
Pillow==7.2.0
pip==20.2.3
protobuf==3.13.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.1.0
setuptools==49.6.0.post20200814
six==1.15.0
tensorboardX==2.1
torch==1.0.0
torchvision==0.2.1
tornado==6.0.4
tqdm==4.49.0
wheel==0.35.1
